back-end/requirements.txt
Flask==3.0.0
flask[async]==3.0.0
setuptools==69.0.2
pymongo===4.6.1
Flask-PyMongo===2.3.0
Flake8==6.0.0
python-dotenv==1.0.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1